Update documentation for class_meshinstace.rst
MeshInstance contains the set(String property, var value) property for setting morph shapes from GDScript, but it isn't documented. I stumbled over a year old reference to it in the doc's of the old site. I've added reference to this function in the member function list and the member function description of this document.