Fix shading language variable
After some testing, it looks like the COLOR variable in canvas_item fragment shaders is write-only, not read only. I'm able to set the COLOR variable using a vec4, and if I sample from the texture, can modify the color of a pixel. Running the code 'COLOR = COLOR;', along with any other modification of COLOR, makes the image entirely white.

Is there any way I could verify that the docs for shader variables are correct? I'd like to learn the shading language but have had a tough time writing basic shaders so far.