Updated commands now that macOS is 64-bit exclusive
Follow-up on this issue: https://github.com/godotengine/godot-docs/issues/921

As scons will compile a ".64" executable by default, not a ".fat" executable, I have updated the commands to reflect that. I also added a section for manually compiling a ".fat" executable for running on a 32-bit machine.

 Updated commands now that macOS is 64-bit exclusive

As scons will compile a ".64" executable by default, not a ".fat" executable, I have updated the commands to reflect that. I also added a section for manually compiling a ".fat" executable for running on a 32-bit machine.

Small Grammer

Removed an erroneous a and a comma.