Links for everything that needs to be installed
I really appreciate the android sdk link, and honestly had never heard of Open JDK or Oracle JDK before today. So a quick google search brought me to the link I've put in this proposed change. If if it's correct, and the placement is good then please accept this PR.