FPS tutorial remaster (#1357)
* Started remastering first two parts of the FPS tutorial.

Now both parts follow a consistant coding style,
have up to date Zip files for every part.

* Finished the first draft for part 2, and did some edits to part 1.

* Finished writing the first draft for part 3. (Still needs editing)

I also added a blank page for part 5 and added it to the index.

* Updated and fixed a few things in Part 3

* Removed the old part 4 of the FPS tutorial and replaced it with a new part!
(Which would have been part 5 had part 4 not been removed)

* Added part 5 of the FPS tutorial.
Made a few changes to parts 4 and 1.
Added a template/overview of what to write in part 6.

* Finished part 6 of the FPS tutorial.
Now every part of the FPS tutorial is written out!

Made some minor edits to part 5.

* Added all of the zip files for the first three parts of the FPS tutorial.
Updated/Redid all of the pictures for the first three parts as well.
Fixed a few minor issues/inconsistencies in the written material.

* Added all of the zip files for the last three parts of the FPS tutorial.
Updated/Redid all of the pictures for the last three parts as well.
Fixed a few minor issues/inconsistencies in the written material.
Updated the blender files zip.

* Did some editing on parts 3 through 5 on the FPS tutorial.