Add .readthedocs.yml file to disable htmlzip and epub builds
Reference: https://docs.readthedocs.io/en/latest/yaml-config.html

`htmlzip` target is too heavy and slows down build a lot (or makes it timeout),
so disabling for now. We could reenable it if an option is given for a non
`singlehtml` target, see https://github.com/rtfd/readthedocs.org/issues/4190.

Also disabling `epub` target for now, as it also makes builds slower and we
iterate too fast on the master branch for it to be meaningful.
We could eventually build ePub versions manually once a month or so and upload
those somewhere if there's demand for it.