Remove tutorial on linear algebra for rotations
Though it covers an interesting topic in a way that can be beneficial
to many Godot users, after discussion with doc team members and reduz
we feel that its current scope is too out of phase with the target
readership of the Godot docs.

This tutorial would work better hosted out of the Godot docs, and
referenced with a `.. seealso::` link for those who want to dig further
into the math meat.

Thanks a lot to tagcup for this contribution though (as well as the
many contributions to the engine's math and rendering), and sorry for
not better describing what kind of "advanced" math tutorial would be
fitting for the Godot docs.

Sphinx: Also reverted from mathjax to imgmath, as the former implied
running JS code on every single page (regardless of their use of LaTeX
formulas to renderer), so for now imgmath is simpler and good enough.