Change comment for `s.queue_free()`
In the previous section, we document `free` as being immediate, and here
we contrast that with `queue_free`. However, each function has the same
comment implying that both `free` and `queue_free` are immediate.

I believe the point of this documentation is to point out that
`queue_free` is _not_ necessarily immediate and will wait until it is
safe to free the node.