FPS Tutorial Part02-04 general phrase corrections (#1560)
* FPS Tutorial Part02 general phrase corrections

* FPS Tutorial Part03 minor phrase corrections

Also I have fixed the audio name convention so now the sound will play as intended.

* FPS Tutorial Part04 minor phrase corrections

* Changes to the joypad input handling  to include the PS controller

I have made some changes so the tutorial works for both the Xbox controller and the Playstation controller. However I have not managed to find a solution to include them in the same script so now it works just with one of them or the other.