FPS Tutorial Part05-06 general phrase corrections
Complete Part05 general phrase corrections

Updated the line 722 in Turret.gd script

Updated a line to be the same as in the current master branch in the Turret.gd script. Before I just pasted my code from the editor following the tutorial from the website witch has an outdated line of code

Removed an empty line inserted by mistake

Part06 partial phrase corrections

Part05 fixed a typo in the corrections

Complete Part06 general phrase corrections

Changed line 488 to be more clear for beginners

I changed the line 488 from "if ray_result:" to "if ray_result != null" so it should be more in line with the code standards and to be more easy to understand by the new users learning how to code.