Removed an unnecessarily controversial line
I am not even going to start a discussion about GDscript vs. C#, but this line is unnecessary at this place and will only lead to false impressions. If you are in a company that intends to use C# you won't need this information. If you don't know about C# yet there's a good chance that you will be better off using GDscript.