Add clarifying paragraph on encryption page (#1672)
I wouldn't want to mislead people into thinking that encrypting and decrypting a save file locally on the player's computer can prevent editing and sharing of save files. It can't. The only way to prevent unauthorized editing and sharing of save data is to keep the save data on a remote server, and also running the game logic on that server instead of locally on the player's computer.