Tweak background_loading
Fix some punctuation and grammar in background_loading. 
Notes: 
"**setup**" (a single [compound] word) is a  _noun_  (a name for an object; a signifier [in Ferdinand de Saussure's semiology], a referrer, an identifier); 
"**set up**" (two separate words) is an 
_idiomatic_  (only makes sense when you take it as it is in its proper context) 
_transitive_  (it takes a [direct] object [in a sentence -- grammar]) 
_phrasal_  (made up of more than one word)
_verb_  (something that expresses an action being done; a process).

Data on which functions are expected to operate:
"**argument(s)**" is the name for what a function takes in when it is being called/invoked/"conjured"/"summoned" [:imp: occult language right there :unamused:]; 
"**parameter(s)**" is the name for that/those factor(s) that alter[s] what a function is designed to return when one is defining it :scroll:, usually introduced/"declared" in the function header.

"Low**-**level" is the correct spelling of the adjective.