Fix punctuation in size_and_anchors
Fix punctuation in size_and_anchors. 
I presume the illustrations are stragglers from an older version of Godot, the look of which I like quite a bit more. 
At any rate, the current version of Godot doesn't present its options in the manner set forth by this page; there seem to have been a number of minor changes done and  ``Grow Direction``  is what it pertains to now.