vertex_array Vector3 correction
`vertex_array` is defined as a Vector3, but it only contains 2 elements (I assume this was a mistake). So I have added the third element (being 0) to complete the size (i.e. make it a 3-component vector). These vertex vectors should now define the coordinates of a triangle in 3D space.