get_skidinfo() misleading definition
It seems (please correct me if you understand it otherwise) that the description for the `get_skidinfo()` function is the inverse of what actually happens. I have run some simple tests, and looks like setting low  `Friction Slip`  (e.g. try 1 or 2) causes more skidding and returns smaller values for `get_skidinfo()`, while when the `Friction Slip` is increased, the car skids less and values printed to the console increase (get closer to 1.0). So it seems that a value of 0 means skidding, while a value of 1.0 means no skidding (the description says the exact opposite from what I understood).