Clarify exporting a PCK from the command-line
The "Runnable" preset is actually used to define which presets
can be run from an one-click deploy in the editor. It is currently
only used for Android and HTML5.

From testing on commit
https://github.com/godotengine/godot/commit/5441aaf768d6dd4c3d8465e6b340ae38ddc7db1d,
the export preset's "Runnable" status has no bearing on whether
a PCK file can be created or not.

Likewise, an executable project can still be exported
from an export preset not marked as "Runnable".