Added C++ flags and 'bin' dir Creation
The build process fails when it tries to find 'demo/bin'. When we create
our demo the bin directory is not created by default, so let just make
it here. The build process also fails on OSX because the project uses
c++11 features. Lets add the c++17 flag so this error no longer happens and
because its used in other platform builds.