Remove redundant adjective in introduction.rst
The adjective "main" before "raison d'etre" in the second bullet of the last section (Organization of the documentation) is redundant. 

"raison d'etre" means "the most important reason for something's existence", so to say "the main raison d'etre" is to say "the main most important reason..."

I know this is pedantic, but I wanted to start contributing so I figured I'd tackle something easy the first time.