Added Instancing in Code Tutorial
The Instancing (continued) page initially only had what I can best describe as "best practices" wording that helps explain instancing conceptually, but uses a lot of comparative language like "one of the best engines out there." and starts off a paragraph with "But" and it has no information on actually instancing using GDScript. I will leave this part alone (for the most part) but I'd love this to be moved elsewhere later on.