Reinsert clamp in exporting docs
Before this change clamping the players position got removed with the following explanation:
"We don't need to clamp the player's position because you can't click outside the screen."

While it is true that the player cannot click outside the window, the player can still click outside the playable area. Because we set Aspect to "keep" in the Project Settings, Godot will create Black borders if the window (or phone screen) doesn't match the desired aspect ratio. A player can now click on either of the borders and because clamp got removed the Player node will move out of the playable area and hide in the black border.

This is why the clamp should not be removed.