Adding information and guidelines about how to properly create a standard CSV translation file (#3048)
* Adding guidlines for escaping double quotes

RFC-4180: "If double-quotes are used to enclose fields, then a double-quote appearing inside a field must be escaped by preceding it with another double quote."
I tried to escape it with backslash without knowing the special standards of CSV, it didn't work and I opened an issue for nothing. I think it might be right to explain that directly in the docs so that no ones lack this information and ask for help about that again.

* Adding more info about the CSV standard

Some information from the CSV standard were lacking, most importantly, how to escape double quotes in a string. An example with escaped double quotes has been added to the existing example table. (Thanks to the discord for the translation)