Expanded class_bool Description with examples
I expanded boolean long description with example of when booleans could be use practice because 

Boolean 

Description:

Boolean built-in type.

was pretty vague. As new to programming person it took me longer than it should have to understand what those are and why I would want to use them. Hopefully this will make it clearer to new users reading documentation