Using AnimatedSprite with a sprite sheet
I updated this tutorial to show how to easily use a sprite sheet with the AnimatedSprite class, including some screen shots.  I changed the old wording that makes it sound like you cannot use a sprite sheet with AnimatedSprite.  I also changed the wording that makes it sound as if you must choose between using the AnimatedSprite class or the AnimationPlayer, as you can use an AnimationPlayer with an AnimatedSprite.