Update 2D animation tutorial to show how to use a sprite sheet with AnimatedSprite (#3159)
* Added images for AnimatedSprite from spritesheet

Added images to show how to use the AnimatedSprite class with a spritesheet in the tutorial.

* Using AnimatedSprite with a sprite sheet

I updated this tutorial to show how to easily use a sprite sheet with the AnimatedSprite class, including some screen shots.  I changed the old wording that makes it sound like you cannot use a sprite sheet with AnimatedSprite.  I also changed the wording that makes it sound as if you must choose between using the AnimatedSprite class or the AnimationPlayer, as you can use an AnimationPlayer with an AnimatedSprite.

* Changed wording on proposed update

I changed the wording to remove "easily" and verbs ending in "ing" to meet the doc's writing guidelines.