Clarification on "Identifier" term
When filling iOS export forms, I got a red error that I'm missing an "Identifier". As an iOS developer and I find the term "Identifier" confusing and it should be explicitly told that it's a Bundle Identifier. When I'm looking at the form itself I thought this doesn't need to be filled as there was already a placeholder with reverse dns notation and I thought it was a default value.