Safe line vs Unsafe line unclear label (#3366)
Although line 6 is safe, line 7, that actually is what the paragraph talking about, is unsafe and line 8 is safe.
The way it was before, it was in reverse order, making it confusing which is safe and which is not. So I changed the order to be as the picture.
To further assure it is clear, I added the line numbers in parentheses.
The parentheses could be excessive verbosity, but the order importance stands. I believe it was confusing before.
Alternatively, the picture could be changed, having the safe line first. Then it's clearer as the unsafe line it's the only one different and the order would be also correct; 2 safe lines, one unsafe for last. Specifying the lines would not be needed then.