Appease our great search engine overlords
This hopefully fixes all translated pages showing up in search engine results, and makes the STABLE version of each page canonical.
In turn, this allows us to re-enable indexing of the version-specific pages (see robots.txt changes), as search engines should prefer the canonical (stable) version, and only show the other versions if no canonical (stable) version exists (i.e. because that feature is only in latest, or was removed in stable).

It adds proper canonical links for all generated pages, and fixes the existing links between the various translations of a page by both ensuring the pages links to itself with the proper language tag, and by properly linking to the full path of other translated versions.

(cherry picked from commits:
- fd5f6f49097ab916ff3f5daa433c6e57d0a59716,
- 263ff562516c98fa6113c920ea7281e21d2d9f79,
- e21df0671fe5a5aeb0e6253084dd842bad6114da,
- 24781e377baad9375aaa7c7175e0bedb5c9c0414,
- 66d185d5d2d51cfd8c28716930d1ca04ff51090b,
- 3c79f3e3218bff53a6303e7b85ebe2089c09c4b6)

Co-authored-by: Rémi Verschelde <rverschelde@gmail.com>