compiling on osx needs to be forced to python3
Since python2 is deprecated, compiling for MacOS needs to be forced to python3.
MacOS has python2 as `python` for it's default env. Changing your python env on OSX is not advised as the system uses it for sereval MacOS features. This simple python3 forcing seems to be the cleanest way to get it working on any MacOS.