Add note about register_types file locations
I organize my projects with 'include' and 'src' directories. When I put the
register_types.[h,cpp] in them, the module would "compile fine" (no errors),
but it wouldn't work.

This took a long time to figure out, so this PR adds a note to ensure that
these files are in the right place to be picked up by the build.