high-level networking: remove bad practice; bomber demo more explicit
Removed an instance of a bad practice (letting the clients pass their
own rpc caller id to the server and having the server trust it), as
well as making the server the only one able to unpause the game.

The first of the two is a fairly harmful habit to teach, especially
in a first introduction.

Also made the bomber demo way more explicit (and verbose) by avoiding
as many ambiguous pronouns as possible, instead being quite explicit
about which nodes or methods are being referred to at any time, and
also some intermingled motivations/reminders for why these steps are
being taken.