Clarify misleading sentence
This is a section for beginners, and until this point,
the writing has been easy and accessible, but the GDScript
Reference is over 6.2k words long and difficult to grok in
a few minutes by a beginner.

This commit changes this sentence to instead suggest skimming
and bookmarking the page for later, because I think this was
the original intent.