Update GDScript examples - Step By Step: Scripting
In Godot 4.0, the connect() function no longer takes an object and
a string name of a function, but rather the new Callable type. Following
the old instructions leads to errors.

Unfortunately, I don't have C# development set up on my machine, so I
have not updated the C# examples as I'm currently unable to validate that
they work as intended.