Document the development philosophy
The documentation lacks a thorough description on general principles
which govern Godot development, and acts as introductory material for
potential contributors which would like to get familiar with the
development aspects that start with "Why?" rather than "How?", as with
the existing "Best practices for engine contributors" page.

This is different to "Design philosophy" because this is mainly targeted
for people who express interest in contributing as potential engine
developers, rather than developers that simply use the engine.