Update vector math doc
In the section explaining about normalizing a vector. In the warning, it says we cannot normalize a vector with a length of 0. And doing so will result in an error.

But after testing out normalizing a vector ( via normalized() function ), no errors are reported. Also printing out to the console shows (0, 0, 0) as the output. Similarly, normalizing manually by dividing it by the length shows strange result if printed out, but still no error.

I think GDScript ( not tested with other languages ) handles normalization of vector in such a way that we don't have to worry about normalizing a vector of length 0 ( only when done through normalized() function ).