Update best_practices_for_engine_contributors
Remove the redundant use of the word "actual" in "The Problem always comes first".  In "The Solution must be discussed with others" Remove an unnecessary comma as well as the redundant use of the word "own", which is not necessary in this context because "their" is used already. This is a minor improvement to conciseness, clarity and readability.