Re-write instancing.rst
This is a complete rewrite and merge of instancing.rst and
instancing_continued.rst. The goal was to simplify the language and
unify the tutorial around the falling ball. Some instructions are more
detailed than in the original article.

There are also more pictures, all new, and using yellow outlines for
color-blind people.

Closes #4175
Closes #4171