Changes the order and text of UI workflows list
When the two UI workflows are introduced, they are presented in a list in the inversed order of the detailed text that follows.
This is confusing for someone who's approaching this topic for the first time. 

This proposal changes the text and the order of the list to be similar to the main titles of the detailed text.