Update exporting string enum to include default value
In Godot v3.2.3.stable.official when exporting a var as a String enum, a default value must be provided or else the variable remains empty when running the scene. Setting a default value of `""` resolves this issue. 

This documentation appears to be for a newer syntax of GDScript though so I don't know if what I've proposed is exactly right.