Rewrite release policy to clarify versioning and include 3.3
We are often asked if Godot follows SemVer, so I rewrote the relevant section
to clarify that our versioning is based on SemVer but takes some liberties to
accommodate the complexity of versioning so many interconnected components with
a single version number. Examples should help make the rationale clear.

Additionally, with the upcoming 3.3 release we're aiming at increasing the
release pace for future 3.x minor releases, and keep patch releases better
focused on actually being maintenance releases (with 3.2.2 being a significant
exception to that rule).