Use a separate GitHub Actions workflow to build offline documentation (#4733)
This is required to get a permalink using https://nightly.link
as we need the workflow containing the offline docs ZIP artifact
to always be the latest one. Creating a second workflow is a good
way to solve this.

This also makes it easier to distinguish between linting checks
and offline documentation builds in the repository's Actions tab.