Marking overlapp area (grey) in 3x3minimal wall tilemap templates
Currently the area I mapped here in grey is was marked yellow. Yellow is supposed to mark an area that shows the top of the wall. This area marked in grey is never shown as top of the wall, because other tiles are suppose to overlap it. I hope marking it in grey and adding a text in the description of the colors makes it more clear how such a Tileset ought to be drawn.