Update compiling_for_osx.rst
Went in and update the compile and build process for both the target = debug and the target = release_debug version.  Before this the compile section contained all the code needed to run the target = debug version while building the applications section had only the code needed for the target = release_debug version.