Update 05.spawning_mobs.rst
Implicit casting of "Godot.Node" to "Mob" is not allowed, so I tried explicit casting and it worked. Please adapt the changes, it took me quite some time to get to the solution, as I thought this was intended and I thought at first, that I did the mistake. Until this point the 3D-Guide worked perfectly fine.