Changed layout to "Top Wide" in final GUI step
Original says to use "Center Top" by mistake. This was confirmed by downloading the finished example project, where the GUI node is properly marked "Top Wide" to produce the intended horizontal sliding of the GUI. "Center Top" causes the GUI to be pinned in the top/center but not scale horizontally with the window.