Optimizing for size: add classes that are disabled by `disable_3d`/`disable_advanced_gui`
Adds a list of classes that are disabled when the explained build options are used.
Generated by parsing `scene/register_scene_types.cpp` and `register_types.cpp` of
all modules.

Also moves the mention of the online build options generator to the section where
`custom.py` is introduced and do some general reformatting and rewording.