Don't disable WebP when compiling small binaries in Optimizing for size
WebP decoding is now required in pretty much any project since Godot 3.4
(unless you force PNG for lossless textures in the Project Settings).

Therefore, it's not a good idea to recommend disabling it for most
people. Not to mention the added binary size will be outweighted by
the smaller lossless WebPs most of the time.