Use "Prefer local solutions" in best practices for engine contributors
Stating that "Solution must be local" is denying the fact that some problems
cannot be solved locally. Therefore, documentation should recommend
contributors to first look at solutions which are closer to the problem first,
and only go for solution that touch core if a problem cannot be solved
locally, for instance.