Fix grammar issue in "Ways to contribute" page
In the section "Filing an issue on GitHub", the subsection "How to reproduce the bug" says "Even if you think that the issue is trivial to reproduce, adding a minimal project that lets reproduce it is a big added value." which is missing a word. Changing it to "...lets everyone reproduce it..." fixes the grammatical error.