Update groups.rst
Hi. I suggested a slight modification to give information right away about creating a new group programmatically, and also how to remove a node (these info were dispersed on forums).

But.. it still lacks a thing:  is it possible/necessary to delete a group? Or, are they automatically disposed of once there are no members left? I fell like the info should be explicitated.