Reverse parenting order of Sprite and AnimationPlayer in tutorial
The docs had a sprite as the child of the player, then *immediately* after
says to not do that. Why mislead the reader at all and not just do it
the better way from the start? There nothing really gained from the
sprite being a child here, not even simplicity.