Streamline bug reporting process
Before this change, there was multiple different bug reporting guides.
This change removes the guide that was in the docs and replaces it with
a note that says to follow [the template][1].

The idea behind this change is to make [the template][1] the only place
bug reporters have to go (i.e., they should no longer have to look at a
guide on another page). This will (hopefully) make reporting issues
easier.

This commit partially implements this proposal:
<godotengine/godot-proposals#4083>

[1]: https://github.com/godotengine/godot/blob/master/.github/ISSUE_TEMPLATE/bug_report.yml