Outdated CS50 Link
CS50 link was to the 2020 course. I have replaced the link to a generic link that will redirect to the most up to date version of the course.

A header on the site for the 2020 link reads:
This is CS50x 2020, an older version of the course. See cs50.harvard.edu/x for the latest!