Improved title of 3d_text.rst for descriptive navigation
Previously, it displayed as "Introduction" under the 3D tab, making it incredibly unclear what the page was actually about. If I understand ReadTheDocs generation of the menu correctly, this will give it a much clearer and more descriptive name of "3D text" in the navigation menu.
I based my exact title on the format of the "3D lights and shadows" documentation title, figuring it was the closest in content type to what is included in the 3D Text documentation.